+++ infra/resources/ced/prod/locals.tf
@@ -19,6 +19,16 @@ locals {
 
   secrets_id_template = "${data.azurerm_key_vault.common.vault_uri}secrets/%s"
 
+  portal_be_container_app_name = provider::dx::resource_name({
+    prefix          = local.prefix
+    environment     = local.env_short
+    location        = local.location
+    domain          = local.domain
+    name            = "portal-be"
+    instance_number = 1
+    resource_type   = "container_app"
+  })
+
   # Portal BE Container App configuration
   portal_be = {
     target_port = 8080
@@ -28,6 +38,9 @@ locals {
     app_settings = {
       APPLICATIONINSIGHTS_ENTRA_ID_AUTH_ENABLED = "false"
       APPINSIGHTS_SAMPLING_PERCENTAGE           = "100"
+      TELEMETRY_SERVICE_NAME                    = local.portal_be_container_app_name
+
+      AZURE_LOG_LEVEL = "error"
 
       PORT            = "8080"
       POSTGRES_HOST   = "${module.postgresql.postgres.name}.postgres.database.azure.com"
@@ -40,6 +53,7 @@ locals {
 
       CED_PORTAL_FE_BASE_URL = "https://${module.portal_fe_static_web_app.custom_domain}"
       CED_PRODUCT_ID         = "prod-ced"
+      ADMIN_FISCAL_CODES     = ""
     }
 
     startup_probe_path   = "/api/info/startup"